requirements.txt
spotipy==2.22.1
customtkinter==5.2.1
pillow==10.1.0